Fixed road type mappings Fixed camel case of names with umlauts and other special characters
