+++ Makefile
@@ -48,7 +48,7 @@ OSMPBF_LIBRARY=-losmpbf
 endif
 
 INCLUDES=${OSMIUM_INCLUDE} -Iplugins ${OSMPBF_INCLUDE}
-LIBS=-lbz2 -lgdal -lexpat -lgeos -lpthread -lz -lprotobuf-lite -lboost_system -lboost_filesystem ${OSMPBF_LIBRARY}
+LIBS=-lbz2 -lgdal -lexpat -lgeos -lpthread -lz -lprotobuf-lite -lboost_system -lboost_filesystem -licuuc ${OSMPBF_LIBRARY}
 
 CXXFLAGS=-std=c++11 
 DEBUG_FLAGS=-O0 -g
